requirements.txt
flask>=2.3.0
pypdf>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.10.0
pymupdf>=1.23.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.25.0
pytesseract>=0.3.10
pdf2image>=1.16.0
